requirements.txt
django
transformers
torch
sentence-transformers
faiss-cpu
openai
python-dotenv
psycopg2